refactor: 폴더명 변경
- api -> apis
