+++ frontend/src/apis/group.ts
@@ -1,4 +1,4 @@
-import { privateGroupsAxiosInstance } from '@/api/axios';
+import { privateGroupsAxiosInstance } from '@/apis/axios';
 import { Group } from '@/types/group';
 
 const getGroups = () => privateGroupsAxiosInstance.get('');
