+++ frontend/src/layouts/NavigationLayout/components/GlobalFootbar/GlobalFootbar.tsx
@@ -3,7 +3,7 @@ import { useParams } from 'react-router-dom';
 import { Group } from '@/types/group';
 import GlobalFootbarGlobalbar from '@/layouts/NavigationLayout/components/GlobalFootbarGlobalbar/GlobalFootbarGlobalbar';
 import GlobalFootbarFootbar from '@/layouts/NavigationLayout/components/GlobalFootbarFootbar/GlobalFootbarFootbar';
-import { getGroups } from '@/api/group';
+import { getGroups } from '@/apis/group';
 
 function GlobalFootbar() {
   const [isLoading, setIsLoading] = useState(true);
