+++ frontend/src/pages/GroupInitPage/GroupInitPage.tsx
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 
 import GroupInitContainer from '@/pages/GroupInitPage/components/GroupInitContainer/GroupInitContainer';
-import { getDefaultGroup } from '@/api/group';
+import { getDefaultGroup } from '@/apis/group';
 import { getLocalStorageItem } from '@/utils/storage';
 import { AxiosError } from 'axios';
