+++ frontend/src/pages/GroupInitPage/components/GroupCreateForm/GroupCreateForm.tsx
@@ -1,6 +1,6 @@
 import { ChangeEvent, FormEvent, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { createGroup } from '@/api/group';
+import { createGroup } from '@/apis/group';
 import { Group } from '@/types/group';
 import GroupCreateFormNameInput from '@/pages/GroupInitPage/components/GroupCreateFormNameInput/GroupCreateFormNameInput';
 import GroupCreateFormSubmitButton from '@/pages/GroupInitPage/components/GroupCreateFormSubmitButton/GroupCreateFormSubmitButton';
