+++ frontend/src/pages/PollProgressPage/components/PollProgressButtons/PollProgressButtons.tsx
@@ -1,6 +1,6 @@
 import { useTheme } from '@emotion/react';
 import { useNavigate } from 'react-router-dom';
-import { deletePoll } from '@/api/poll';
+import { deletePoll } from '@/apis/poll';
 import { Poll, getPollResponse } from '@/types/poll';
 import Button from '@/components/Button/Button';
 import FlexContainer from '@/components/FlexContainer/FlexContainer';
