+++ frontend/src/pages/RoleMainPage/RoleMainPage.tsx
@@ -2,7 +2,7 @@ import { StyledContainer } from '@/pages/RoleMainPage/RoleMainPage.styles';
 import RoleMainHeader from '@/pages/RoleMainPage/components/RoleMainHeader/RoleMainHeader';
 import RoleMainProgress from '@/pages/RoleMainPage/components/RoleMainProgress/RoleMainProgress';
 import RoleMainResult from '@/pages/RoleMainPage/components/RoleMainResult/RoleMainResult';
-import { getRolesHistories } from '@/api/role';
+import { getRolesHistories } from '@/apis/role';
 import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { Group } from '@/types/group';
