+++ frontend/src/routes/PrivateRoute.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
 import { Navigate, Outlet } from 'react-router-dom';
-import { getUser } from '@/api/auth';
+import { getUser } from '@/apis/auth';
 import useAuthDispatchContext from '@/hooks/useAuthDispatchContext';
 import { getLocalStorageItem } from '@/utils/storage';
